i added from date and to date in handlers for fetching the logs and
